fix clang-tidy warning and use 64 bit ints for sizes.

+++ source/server/admin/stats_request.cc
@@ -42,7 +42,7 @@ bool StatsRequest::nextChunk(Buffer::Instance& response) {
 
   // nextChunk's contract is to add up to chunk_size_ additional bytes. The
   // caller is not required to drain the bytes after each call to nextChunk.
-  uint32_t starting_response_length = response.length();
+  uint64_t starting_response_length = response.length();
   while (response.length() - starting_response_length < chunk_size_) {
     while (stat_map_.empty()) {
       switch (phase_) {
